moved vmtemplate server
